Fixed to actually tesselate in the described manner
Realized the problem described tesselating in a totally different way that I interpreted it.

If r is the rotation angle, I assumed that tiles in the first row would be rotated 0r, 1r, 0r, 1r.. and so on times while tiles in the second row would be rotated 3r, 2r, 3r, 2r... and so on. As if the tiles were being rotated around a point in the center of every 4 tiles.

I realize now that the solution is looking for the tiles in the first row to be rotated 0r, 1r, 2r, 3r... and so on, while tiles in the second row are to be rotated 1r, 2r, 3r, 4r (or 0r)... and so on, repeating in this pattern until the full width and height is fleshed out.

I've gone and reimplemented the solution to accomodate this. And to be honest, it's much cleaner now than it was before.